requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
annotated-types==0.7.0
async-timeout==4.0.3
attrs==24.2.0
bitarray==3.0.0
certifi==2024.8.30
charset-normalizer==3.4.0
ckzg==2.0.1
cytoolz==1.0.0
eth-abi==5.1.0
eth-account==0.13.4
eth-hash==0.7.0
eth-keyfile==0.8.1
eth-keys==0.6.0
eth-rlp==2.1.0
eth-typing==5.0.1
eth-utils==5.1.0
exceptiongroup==1.2.2
frozenlist==1.5.0
hexbytes==1.2.1
idna==3.10
iniconfig==2.0.0
multidict==6.1.0
packaging==24.1
parsimonious==0.10.0
pluggy==1.5.0
propcache==0.2.0
pycryptodome==3.21.0
pydantic==2.9.2
pydantic-core==2.23.4
pytest==8.3.3
python-dotenv==1.0.1
pyunormalize==16.0.0
regex==2024.9.11
requests==2.32.3
rlp==4.0.1
tomli==2.0.2
toolz==1.0.0
types-requests==2.32.0.20241016
typing-extensions==4.12.2
urllib3==2.2.3
web3==7.4.0
websockets==13.1
yarl==1.15.2
